refactor: split TokenBalancesSection + OctavPortfolioPanel into smaller components
Per review feedback, breaks up the two largest files in the diff into
smaller, mostly-stateless components and extracts pure helpers into
app/utils. State and effects stay in the parent orchestrators; the new
components only render and raise events.

Parents:
- OctavPortfolioPanel.tsx: 483 → 190 lines (-60%)
- TokenBalancesSection.tsx: 841 → 662 lines (-21%)

New utils:
- format.ts            — formatUsd, formatBalance, shortAddr, readableText
- octav-actions.ts     — actionsForPositionTypes
- erc20-abi.ts         — ERC20_ABI

New components:
- OctavAssetListRow    — single asset row in a position's asset list
- OctavPositionGroup   — one position inside an expanded protocol card
- OctavProtocolCard    — collapsible protocol card (owns only its open state)
- OctavChainFilterBar  — the chain-filter strip + hint + reset link
- OctavCopyJsonButton  — the Copy JSON button with its transient pill
- TokensAddTokenForm   — the "+ Add Token" input row
- TokensTable          — the balances table (header + rows)
- TokensJsonEditorModal — the bulk JSON editor modal

No behavior changes. tsc + eslint + prettier clean.
